+++ test/app/Main.hs
@@ -16,9 +16,10 @@ import Game.Werewolf.Test.Game
 import Game.Werewolf.Test.Player
 
 import Test.Tasty
+import Test.Tasty.QuickCheck
 
 main :: IO ()
-main = defaultMain =<< tests
+main = defaultMain . localOption (QuickCheckTests 20) =<< tests
 
 tests :: IO TestTree
 tests = return . testGroup "Tests" $ concat